Return SolrServers to its original state. So newly introduces SolrServerElements, returns to be SolrServers This way backward compatiblity is at its fullest.
netstandard 2.0 users will simply use new List<ISolrServer>
